fix: sandbox alert structure and style
